requirements.txt
streamlit
easyocr
numpy
Pillow
gspread
google-api-python-client
google-auth
opencv-python-headless